docs: clarify Icon() for templ and other template system
I extended the examples to include how to use Icon() in other non-`html/template` templating systems.

**Rationale:**

I missed this functionality on my first pass of the readme. It was in the API section but not the examples, making it easy to overlook. This is likely a common use-case for non-`html/template` users so should be more prominent.